TeleMusickAnywhere-main/requirements.txt
pyTelegramBotAPI==4.4.0
Pillow
qrcode
opencv-python-headless
sqlalchemy
SpeechRecognition